trx_energy_rental/requirements.txt
# TRX能量租赁系统依赖

# Web框架
Flask==2.0.1
Flask-WTF==0.15.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0

# 数据库
SQLAlchemy==1.4.23
pymysql==1.0.2
alembic==1.7.1

# TRON区块链相关
tronpy==0.2.1
tron-api-python==3.1.6

# Telegram机器人
python-telegram-bot==13.7

# 其他工具
python-dotenv==0.19.0
requests==2.26.0
gunicorn==20.1.0
schedule==1.1.0
pydantic==1.8.2
cryptography==3.4.8 